Adjust tests for Field-Type
