+++ lib/chip-types/field-type.js
@@ -1,6 +1,5 @@
 var Promise = require("bluebird");
 
-var ChipManager = require.main.require("lib/chip-types/chip-manager.js");
 var Errors = require.main.require("lib/response/error.js");
 
 var default_methods = require("./field-type-default-methods.js");
@@ -33,6 +32,7 @@ function wrap_method_in_promise (declaration, method_name) {
 }
 
 var FieldType = function(declaration){
+	const ChipManager = require.main.require("lib/chip-types/chip-manager.js"); // for some reason it doesn't work when placed at the beginning
 	var self;
 
 	if (declaration instanceof FieldType){
